requirements.txt
pandas==1.5.3
streamlit==1.22.0
openai==0.27.8
transformers==4.30.2
torch==2.0.1
python-dotenv==1.0.0